requirements.txt
APScheduler==3.10.1
mysql-connector-python
pandas==1.3.4
python-decouple==3.8
